we cannot cast a long long as pointer on big endian systems. this is in fact zero on 32 bit systems
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@38910 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/dhcpv6/config.h
@@ -261,7 +261,7 @@ struct cf_list {
 	int line;		/* the line number of the config file */
 
 	/* type dependent values: */
-	long long num;
+	long num;
 	struct cf_list *list;
 	void *ptr;
 };
@@ -305,7 +305,7 @@ extern struct dhcp6_list nisplist;
 extern struct dhcp6_list nispnamelist;
 extern struct dhcp6_list bcmcslist;
 extern struct dhcp6_list bcmcsnamelist;
-extern long long optrefreshtime;
+extern long optrefreshtime;
 
 extern struct dhcp6_if *ifinit __P((char *));
 extern int ifreset __P((struct dhcp6_if *));
